requirements.txt
aenum==3.1.15
aiohttp==3.9.5
aiosignal==1.2.0
annotated-types==0.6.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.1.0
async-timeout==4.0.3
attrs==23.2.0
Authlib==1.3.1
autobahn==22.7.1
backcall==0.2.0
bcrypt==4.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.7.0
blis==0.7.11
catalogue==2.0.10
certifi==2022.9.14
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
contourpy==1.0.6
cryptography==38.0.3
cycler==0.11.0
cymem==2.0.8
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==1.2.0
fastapi==0.111.1
fastapi-cli==0.0.4
fastjsonschema==2.16.2
filelock==3.13.1
Flask==3.0.2
fonttools==4.38.0
frozenlist==1.3.1
fsspec==2023.10.0
future==1.0.0
gensim==4.3.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
hyperlink==21.0.0
idna==3.3
imbalanced-learn==0.11.0
imblearn==0.0
ipykernel==6.17.1
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.17.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.23.2
jupyter-server-proxy==3.2.2
jupyter_client==7.4.5
jupyter_core==5.0.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
langcodes==3.3.0
line-bot-sdk==3.11.0
lxml==5.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.4
mpmath==1.3.0
multidict==6.0.2
murmurhash==1.0.10
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.2.1
nltk==3.8.1
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.4
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==21.3
paho-mqtt==2.0.0
pandas==2.1.2
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
patsy==0.5.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.4
preshed==3.0.9
prometheus-client==0.15.0
prompt-toolkit==3.0.32
psutil==5.9.8
pure-eval==0.2.2
pyasn1==0.6.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pygame==2.6.1
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2023.3.post1
pywin32==305
pywinpty==2.0.9
PyYAML==6.0.1
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
regex==2023.10.3
requests==2.31.0
rich==13.7.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.3
selenium==4.20.0
Send2Trash==1.8.0
shellingham==1.5.4
simpervisor==0.4
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.31
srsly==2.4.8
stack-data==0.6.1
starlette==0.37.2
statsmodels==0.14.0
sympy==1.13.1
termcolor==2.2.0
terminado==0.17.0
thinc==8.2.1
threadpoolctl==3.2.0
tiktoken==0.9.0
tinycss2==1.2.1
torch==2.6.0
tornado==6.2
tqdm==4.66.1
traitlets==5.5.0
trio==0.25.0
trio-websocket==0.11.1
txaio==22.2.1
typer==0.12.3
typing_extensions==4.11.0
tzdata==2023.3
urllib3==2.2.1
uv==0.7.8
uvicorn==0.30.3
vpython==7.6.4
wasabi==1.1.2
watchfiles==0.22.0
wcwidth==0.2.5
weasel==0.3.4
webencodings==0.5.1
websocket-client==1.4.2
websockets==12.0
Werkzeug==3.0.1
widgetsnbextension==4.0.3
wordcloud==1.9.2
wrapt==1.16.0
wsproto==1.2.0
xlwt==1.3.0
yarl==1.8.1
pynvml==11.5.0
